requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
python-jose
passlib